+++ FurniFusion(E-Commerce)/Controllers/User/PhoneController.cs
@@ -6,7 +6,7 @@ using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using System.Security.Claims;
 
-namespace FurniFusion.Controllers.Profile
+namespace FurniFusion.Controllers.user
 {
     [Authorize]
     [ApiController]
