+++ FurniFusion(E-Commerce)/Interfaces/IWishlistService.cs
@@ -4,9 +4,14 @@ namespace FurniFusion.Interfaces
 {
     public interface IWishlistService
     {
+        Task<Wishlist> CreateWishlistAsync(Wishlist wishlist);
         Task<ServiceResult<Wishlist?>> GetWishlistAsync(string userId);
-        Task<ServiceResult<Wishlist?>> CreateWishlistAsync(Wishlist wishlist);
 
-        Task<ServiceResult<Wishlist?>> DeleteWishlistAsync(string userId);
+        Task<ServiceResult<WishlistItem?>> AddWishlistItemAsync(int productId, string userId);
+
+        Task<ServiceResult<WishlistItem?>> RemoveWishlistItemAsync(int productId, string userId);
+
+        Task<ServiceResult<bool>> RemoveAllWishlistItemsAsync(string userId);
+
     }
 }
