+++ FurniFusion(E-Commerce)/Services/WishlistService.cs
@@ -9,38 +9,18 @@ namespace FurniFusion.Services
     {
         private readonly FurniFusionDbContext _context;
 
-
         public WishlistService(FurniFusionDbContext context)
         {
             _context = context;
         }
 
-        public async Task<ServiceResult<Wishlist?>> CreateWishlistAsync(Wishlist wishlist)
+        public async Task<Wishlist> CreateWishlistAsync(Wishlist wishlist)
         {
-            var wishlistExists = await _context.Wishlists.FirstOrDefaultAsync(w => w.UserId == wishlist.UserId);
-
-            if (wishlistExists != null)
-                return ServiceResult<Wishlist?>.ErrorResult("Wishlist already exists", StatusCodes.Status400BadRequest);
-
             var createdWishlist = await _context.Wishlists.AddAsync(wishlist);
 
             await _context.SaveChangesAsync();
 
-            return ServiceResult<Wishlist?>.SuccessResult(createdWishlist.Entity, "Wishlist created successfully", StatusCodes.Status201Created);
-        }
-
-        public async Task<ServiceResult<Wishlist?>> DeleteWishlistAsync(string userId)
-        {
-            var wishlist = await _context.Wishlists.FirstOrDefaultAsync(w => w.UserId == userId);
-
-            if (wishlist == null)
-                return ServiceResult<Wishlist?>.ErrorResult("Wishlist not found", StatusCodes.Status404NotFound);
-
-            _context.Wishlists.Remove(wishlist);
-
-            await _context.SaveChangesAsync();
-
-            return ServiceResult<Wishlist?>.SuccessResult(null, "Wishlist deleted successfully", StatusCodes.Status200OK);
+            return createdWishlist.Entity;
 
         }
 
@@ -51,10 +31,63 @@ namespace FurniFusion.Services
                 .ThenInclude(w => w.Product)
                 .FirstOrDefaultAsync(w => w.UserId == userId);
 
-            if (wishlist == null)
-                return await CreateWishlistAsync(new Wishlist { UserId = userId, UpdatedAt = DateTime.Now });
-
             return ServiceResult<Wishlist?>.SuccessResult(wishlist, "Wishlist retrieved successfully", StatusCodes.Status200OK);
+        }
+
+        public async Task<ServiceResult<WishlistItem?>> AddWishlistItemAsync(int productId, string userId)
+        {
+            var userWishlist = await GetWishlistAsync(userId);
+
+            var existingWishlistItem = userWishlist.Data!.WishlistItems.FirstOrDefault(w => w.ProductId == productId);
+
+            if (existingWishlistItem != null)
+                return ServiceResult<WishlistItem?>.ErrorResult("Wishlist item already exists", StatusCodes.Status400BadRequest);
+
+            var wishlistItem = new WishlistItem
+            {
+                WishlistId = userWishlist.Data!.WishlistId,
+                ProductId = productId,
+                CreatedAt = DateTime.Now
+            };
+
+            var createdWishlistItem = await _context.WishlistItems.AddAsync(wishlistItem);
+
+            await _context.SaveChangesAsync();
+
+            return ServiceResult<WishlistItem?>.SuccessResult(createdWishlistItem.Entity, "Wishlist item added successfully", StatusCodes.Status201Created);
+
+
+        }
+
+        public async Task<ServiceResult<bool>> RemoveAllWishlistItemsAsync(string userId)
+        {
+            var userWishlist = await GetWishlistAsync(userId);
+
+            if (userWishlist.Data!.WishlistItems.Count == 0)
+                return ServiceResult<bool>.ErrorResult("Wishlist is already empty", StatusCodes.Status404NotFound);
+
+            _context.WishlistItems.RemoveRange(userWishlist.Data!.WishlistItems);
+
+            await _context.SaveChangesAsync();
+
+            return ServiceResult<bool>.SuccessResult(true, "Wishlist items removed successfully", StatusCodes.Status200OK);
+        }
+
+        public async Task<ServiceResult<WishlistItem?>> RemoveWishlistItemAsync(int productId, string userId)
+        {
+            var userWishlist = await GetWishlistAsync(userId);
+
+            var existingWishlistItem = userWishlist.Data!.WishlistItems.FirstOrDefault(w => w.ProductId == productId);
+
+            if (existingWishlistItem == null)
+                return ServiceResult<WishlistItem?>.ErrorResult("Wishlist item not found", StatusCodes.Status404NotFound);
+
+            _context.WishlistItems.Remove(existingWishlistItem);
+
+            await _context.SaveChangesAsync();
+
+            return ServiceResult<WishlistItem?>.SuccessResult(null, "Wishlist item removed successfully", StatusCodes.Status200OK);
+
         }
     }
 }
